refactor(common-ui): share the expo-updates availability guard
The guard added with the development-build fix existed three times: once in
apps/frontend and once as a copied ExpoUpdatesHelper in geonexia and
score-tracker. All three apps already depend on repo-depkit-common-ui, which
already lists expo-updates as a peer dependency, so the check belongs there.

areExpoUpdatesAvailable() and isExpoUpdatesUnavailableError() now live in
packages/common-ui and are exported from its barrel; the app-local copies are
deleted. apps/frontend keeps a thin wrapper in DeviceRuntimeHelper that adds
its own isInExpoGo() detection on top of the shared check.

The unit test moves along with the helper. It stays in the frontend app,
which is the only workspace with a jest setup that can load the code, and
addresses the implementation by path because the common-ui barrel pulls in
native UI components that jest cannot load.
